+++ src/pages/jeongminlee/List/CoffeeCard.js
@@ -1,25 +1,26 @@
 import React from 'react';
+import { useEffect, useState } from 'react';
 // import COFFEE_LIST from '/public/data/listData.json';
 import './CoffeeCard.scss';
 
-function CoffeeCard({ coffeeList }) {
-  // const [coffeeList, setCoffeeList] = useState([]);
+function CoffeeCard() {
+  const [coffeeList, setCoffeeList] = useState([]);
 
-  // useEffect(() => {
-  //   fetch('/data/listData.json')
-  //     .then(res => res.json())
-  //     .then(data => {
-  //       setCoffeeList(data);
-  //     });
-  // }, []);
+  useEffect(() => {
+    fetch('/data/listData.json')
+      .then(res => res.json())
+      .then(data => {
+        setCoffeeList(data);
+      });
+  }, []);
 
-  // const first = coffeeList.filter(el => el.id <= 10);
-  // const BrewedCoffee = coffeeList.filter(el => el.id > 10);
+  const coldBrewCoffee = coffeeList.filter(el => el.id <= 10);
+  const BrewedCoffee = coffeeList.filter(el => el.id > 10);
 
   return (
     <div className="CoffeeCard">
       <div className="menu">
-        {coffeeList.map(el => (
+        {coldBrewCoffee.map(el => (
           <div key={el.id} className="explain">
             <img
               className="cold_brew_coffee"
@@ -32,9 +33,19 @@ function CoffeeCard({ coffeeList }) {
       </div>
       {/* 콜드브루커피 메뉴 */}
       <br />
+      {/* 브루드커피 네비게이션 바 */}
+      <div className="kind">
+        <span className="kind1">브루드 커피&nbsp;&nbsp;</span>
+        <i className="fa-solid fa-mug-hot" />
+        <span>&nbsp;&nbsp;</span>
+        <span className="explain1_list">
+          디카페인 에스프레소 샷 추가 가능(일부 음료 제외)
+        </span>
+      </div>
+      {/* 브루드커피 네비게이션 바 */}
       {/* 브루드커피 메뉴 */}
-      {/* <div className="menu2">
-        {coffeeList.map(el => (
+      <div className="menu2">
+        {BrewedCoffee.map(el => (
           <div key={el.id} className="explain">
             <img
               className="cold_brew_coffee"
@@ -45,7 +56,7 @@ function CoffeeCard({ coffeeList }) {
           </div>
         ))}
         {/* 브루드커피 메뉴 */}
-      {/* 브루드커피 여백 채우기용}
+        {/* 브루드커피 여백 채우기용*/}
         <div className="hidden">
           <img
             className="cold_brew_coffee"
@@ -63,8 +74,8 @@ function CoffeeCard({ coffeeList }) {
           />
           <span className="coffee_name" />
         </div>
-        {/* 브루드커피 여백 채우기용}
-      </div> */}
+        {/* 브루드커피 여백 채우기용*/}
+      </div>
     </div>
   );
 }
