+++ src/pages/jeongminlee/List/List.js
@@ -8,9 +8,6 @@ import { useEffect, useState } from 'react';
 function List() {
   const [coffeeList, setCoffeeList] = useState([]);
 
-  const coldBrewCoffee = coffeeList.filter(el => el.id <= 10);
-  const BrewedCoffee = coffeeList.filter(el => el.id > 10);
-
   useEffect(() => {
     fetch('/data/listData.json')
       .then(res => res.json())
@@ -19,7 +16,6 @@ function List() {
       });
   }, []);
 
-  console.log(coffeeList.id);
   return (
     <div className="body_list">
       <div className="list">
@@ -34,24 +30,7 @@ function List() {
           </span>
           {/* 콜드브루커피 네비게이션 바 */}
         </div>
-        <CoffeeCard
-          coffeeList={BrewedCoffee}
-          // BrewedCoffee={BrewedCoffee}
-        />
-        {/* 브루드커피 네비게이션 바 */}
-        <div className="kind">
-          <span className="kind1">브루드 커피&nbsp;&nbsp;</span>
-          <i className="fa-solid fa-mug-hot" />
-          <span>&nbsp;&nbsp;</span>
-          <span className="explain1_list">
-            디카페인 에스프레소 샷 추가 가능(일부 음료 제외)
-          </span>
-        </div>
-        {/* 브루드커피 네비게이션 바 */}
-        <CoffeeCard
-          coffeeList={BrewedCoffee}
-          // BrewedCoffee={BrewedCoffee}
-        />
+        <CoffeeCard />
       </div>
     </div>
   );
